Refactor: 상품 URI 수정 및 컨트롤러 분리
