+++ src/main/java/com/nhnacademy/marketgg/client/repository/product/ProductAdapter.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.marketgg.client.repository.impl;
+package com.nhnacademy.marketgg.client.repository.product;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -9,7 +9,7 @@ import com.nhnacademy.marketgg.client.dto.response.DefaultPageResult;
 import com.nhnacademy.marketgg.client.dto.response.ProductResponse;
 import com.nhnacademy.marketgg.client.dto.response.SearchProductResponse;
 import com.nhnacademy.marketgg.client.dto.response.common.SingleResponse;
-import com.nhnacademy.marketgg.client.repository.ProductRepository;
+import com.nhnacademy.marketgg.client.repository.product.ProductRepository;
 import java.io.IOException;
 import java.util.List;
 import java.util.Objects;
@@ -23,6 +23,8 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -59,7 +61,7 @@ public class ProductAdapter implements ProductRepository {
 
         HttpEntity<Void> request = new HttpEntity<>(headers);
         ResponseEntity<DefaultPageResult<ProductResponse>> response =
-                this.restTemplate.exchange(gatewayIp + ADMIN_DEFAULT_PRODUCT,
+                this.restTemplate.exchange(gatewayIp + DEFAULT_PRODUCT,
                                            HttpMethod.GET,
                                            request,
                                            new ParameterizedTypeReference<>() {
@@ -76,7 +78,7 @@ public class ProductAdapter implements ProductRepository {
 
         HttpEntity<Void> httpEntity = new HttpEntity<>(headers);
         ResponseEntity<SingleResponse<ProductResponse>> response =
-                this.restTemplate.exchange(gatewayIp + ADMIN_DEFAULT_PRODUCT + "/" + productId,
+                this.restTemplate.exchange(gatewayIp + DEFAULT_PRODUCT + "/" + productId,
                                            HttpMethod.GET,
                                            httpEntity,
                                            new ParameterizedTypeReference<>() {
