+++ src/main/java/com/nhnacademy/marketgg/client/repository/product/ProductRepository.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.marketgg.client.repository;
+package com.nhnacademy.marketgg.client.repository.product;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.nhnacademy.marketgg.client.dto.request.ProductCreateRequest;
