+++ src/main/java/com/nhnacademy/marketgg/client/web/product/ReviewController.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.marketgg.client.web;
+package com.nhnacademy.marketgg.client.web.product;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.nhnacademy.marketgg.client.annotation.Auth;
