+++ src/test/java/com/nhnacademy/marketgg/client/service/impl/DefaultReviewServiceTest.java
@@ -13,7 +13,7 @@ import com.nhnacademy.marketgg.client.dto.request.ReviewCreateRequest;
 import com.nhnacademy.marketgg.client.dto.request.ReviewUpdateRequest;
 import com.nhnacademy.marketgg.client.dto.response.ReviewResponse;
 import com.nhnacademy.marketgg.client.dummy.Dummy;
-import com.nhnacademy.marketgg.client.repository.ReviewRepository;
+import com.nhnacademy.marketgg.client.repository.product.ReviewRepository;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
